refactor: move back components which need own routes
