+++ src/playground/layout/smart-home/app-routing.module.ts
@@ -7,7 +7,7 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
-import { AppComponent } from './components/app.component';
+import { AppComponent } from './app.component';
 
 export const routes: Routes = [
   {
