+++ src/playground/layout/smart-home/app.component.ts
